types: ConsensusParams: add feedback from @ebuchman and @melekes
